chore(cardwired): move Mode enum to it's own file
