+++ crates/cardwire-daemon/src/file/config.rs
@@ -1,7 +1,7 @@
 //! helper to manage cardwired configs, include the user config .toml, and the .json states like
 //! gpu, mode or pci
 use crate::{
-    file::common::{FileKind, create_default_file}, interface::Modes
+    file::common::{FileKind, create_default_file}, types::Modes
 };
 use anyhow::Context;
 use log::warn;
